6.5 actually fixed this time

+++ server.py
@@ -10,9 +10,12 @@ import psycopg2
 
 app = Flask(__name__)
 
-players_db = os.environ['DATABASE_URL']
-content_db = os.environ['HEROKU_POSTGRESQL_JADE_URL']
+# players_db = os.environ['DATABASE_URL']
+# content_db = os.environ['HEROKU_POSTGRESQL_JADE_URL']
 
+
+players_db = "postgres://ccybprhsnhfxgx:b2ce8114ee71abcadd147b42a47b1caa4ae540ac946317a821d04098d40d4acf@ec2-107-20-153-39.compute-1.amazonaws.com:5432/d4b7p3avace43e"
+content_db = "postgres://hbwubwppwmylry:07310537702b404ca848a5c28e8f791fa4bb8abd942f81b8166ea505bf811620@ec2-34-225-103-117.compute-1.amazonaws.com:5432/deml8rlqer0tim"
 
 @app.route('/')
 def sayHello():
@@ -328,15 +331,21 @@ def update_player(player_id):
 def remove_player(player_id):
   with psycopg2.connect(players_db, sslmode='require') as conn:
     cursor = conn.cursor()
-
-    query = "SELECT characters_id FROM characters WHERE player_id=%s"
+    query = "SELECT * FROM players WHERE players_id=%s;" #make sure player exists
     cursor.execute(query, (player_id,))
-    for key in cursor:
-      deletePlayerCharacter(player_id, key)
-
-    query = "SELECT * FROM players WHERE players_id=%s;"
-    cursor.execute(query, (player_id,))
-    if cursor.rowcount > 0:
+    if cursor.rowcount > 0: #if player exists, proceed to delet characters and attributes
+      query = "SELECT characters_id FROM characters WHERE player_id=%s"
+      cursor.execute(query, (player_id,))
+      if cursor.rowcount > 0:
+        for key in cursor:
+          deletePlayerCharacter(player_id, key)
+      query = "SELECT players_attributes_id FROM players_attributes WHERE player_id=%s;"
+      cursor.execute(query, (player_id,))
+      if cursor.rowcount > 0:
+        records = cursor.fetchall()
+        for key in records:
+          query = "DELETE FROM players_attributes WHERE players_attributes_id=%s;"
+          cursor.execute(query, (key,))
       query = "DELETE FROM players WHERE players_id=%s;" #delete player
       cursor.execute(query, (player_id,))
       if cursor.rowcount == 0:
@@ -460,10 +469,10 @@ def delete_room(game_id, room_id):
 def after_requestuest(response):
   return response
 
-# app.debug = True
-# host = os.environ.get('OP', '0.0.0.0')
-# port = int(os.environ.get('PORT', 8080))
-# app.run(host=host, port=port)
+app.debug = True
+host = os.environ.get('OP', '0.0.0.0')
+port = int(os.environ.get('PORT', 8080))
+app.run(host=host, port=port)
 
-if __name__ == '__main__':
-    app.run(threaded=True, port=5000)
+# if __name__ == '__main__':
+#     app.run(threaded=True, port=5000)
